long

+++ src/main/java/com/rentamaquina/maquinaria/app/entities/custom/CountClient.java
@@ -12,23 +12,23 @@ import com.rentamaquina.maquinaria.app.entities.Client;
  * @author daan_
  */
 public class CountClient {
-    private Integer total;
+    private Long total;
     private Client client;
 
     public CountClient() {
     }
 
     
-    public CountClient(Integer total, Client client) {
+    public CountClient(Long total, Client client) {
         this.total = total;
         this.client = client;
     }
 
-    public Integer getTotal() {
+    public Long getTotal() {
         return total;
     }
 
-    public void setTotal(Integer total) {
+    public void setTotal(Long total) {
         this.total = total;
     }
