+++ src/main/java/com/rentamaquina/maquinaria/app/repositories/ReservationRepository.java
@@ -40,7 +40,7 @@ public class ReservationRepository {
         List<CountClient> res=new ArrayList<>();
         List<Object[]> report=reservationCrudRepository.countTotalReservationByClient();
         for(int i=0;i<report.size();i++){
-            res.add(new CountClient((Integer) report.get(i)[1],(Client)report.get(i)[0] ));
+            res.add(new CountClient((Long) report.get(i)[1],(Client)report.get(i)[0] ));
         }
         return res;
     }
